+++ frontend/src/app/page.tsx
@@ -1,13 +1,26 @@
 "use client";
-// import PetsCard from "@/components/adoption-card";
+import PetsCard from "@/components/main_page/adoption-card";
 import Hero from "@/components/home/hero";
+import {Cards } from "@/lib/data";
+
+interface Card {
+  id: number;
+  image: "";
+  name: string;
+}
+
+interface PetsCardProps {
+  Cards: Card[];
+}
 
 export default function Home() {
   const App: React.FC = () => {
     return (
       <section>
         <Hero />
-        {/* <PetsCard /> */}
+        {Cards.map((card) => (
+        <PetsCard image={card.image} name={card.name} /> 
+      ))}
       </section>
     );
   };
